+++ common/src/main/java/net/xolt/freecam/mixins/LocalPlayerMixin.java
@@ -1,6 +1,7 @@
 package net.xolt.freecam.mixins;
 
 import net.xolt.freecam.Freecam;
+import net.xolt.freecam.config.ModConfig;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -22,17 +23,19 @@ public class LocalPlayerMixin {
         }
     }
 
-    // Makes arm movement depend upon FreeCamera movement rather than player movement.
+    // Makes rotation depend upon FreeCamera rather than the player.
     @Inject(method = "getViewXRot", at = @At("HEAD"), cancellable = true)
     private void onGetViewXRot(float partialTick, CallbackInfoReturnable<Float> cir) {
-        if (Freecam.isEnabled())
+        if (Freecam.isEnabled() && !Freecam.isPlayerControlEnabled() && !ModConfig.get().allowInteractionsFromPlayer()) {
             cir.setReturnValue(Freecam.getFreeCamera().getViewXRot(partialTick));
+        }
     }
 
-    // Makes arm movement depend upon FreeCamera movement rather than player movement.
+    // Makes rotation depend upon FreeCamera rather than the player.
     @Inject(method = "getViewYRot", at = @At("HEAD"), cancellable = true)
     private void onGetViewYRot(float partialTick, CallbackInfoReturnable<Float> cir) {
-        if (Freecam.isEnabled())
+        if (Freecam.isEnabled() && !Freecam.isPlayerControlEnabled() && !ModConfig.get().allowInteractionsFromPlayer()) {
             cir.setReturnValue(Freecam.getFreeCamera().getViewYRot(partialTick));
+        }
     }
 }